study_tutor/requirements.txt
streamlit==1.32.0
openai==1.14.0
python-dotenv==1.0.1
requests==2.31.0
PyPDF2==3.0.1
